+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix toggle button styling
 - Miscellanous code improvements
 - Improve container index table
+- Fix bug with ammo not updating after deleting shot group
 
 # v0.7.1
 - Add shading to table component